Fix tMOVCC_pseudo for libpng -O0.
Previous fix to tMOVCC_pseudo will cause a regression on
libpng when -O0 is specified.

Some non live-ins physical registers will be splitted as
well, so we should create virtual registers for them.

This CL implements a complete live variable analysis to
analyze the usage of physical registers among the splitted
machine instructions.

Change-Id: I1974111c3446f0b66e012cead7bdb5dffe4f8b3b